Switched nearly every character pointer in KMMessage and KMMsgPart to QString. This should improve performance and usability.
svn path=/trunk/kdenetwork/kmail/; revision=2217

+++ kmmsginfo.cpp
@@ -206,7 +206,7 @@ const char* KMMsgInfo::dateStr(void) const
 void KMMsgInfo::fromString(const char* aStr)
 {
   char st;
-  unsigned int idate;
+  unsigned long idate;
 
   assert(aStr != NULL);
